renamed gridappsd_mysql.sql to gridappsd_mysql_dump.sql

+++ run.sh
@@ -44,7 +44,7 @@ clean_up () {
 
 viz_url="http://localhost:8080/ieee8500"
 blazegraph_url="http://localhost:8889/bigdata/"
-mysql_file="gridappsd_mysql.sql"
+mysql_file="gridappsd_mysql_dump.sql"
 data_dir="dumps"
 # set the default tag for the gridappsd and viz containers
 GRIDAPPSD_TAG=''
@@ -81,9 +81,11 @@ if [ ! -f "$data_dir/$mysql_file" ]; then
   echo " "
   echo "Downloading mysql data"
   curl -s -o "$data_dir/$mysql_file" "https://raw.githubusercontent.com/GRIDAPPSD/Bootstrap/master/$mysql_file"
-  sed -i.bak -e "s/'gridappsd'@'localhost'/'gridappsd'@'%'/g" $data_dir/$mysql_file
-  # clean up 
-  rm $data_dir/${mysql_file}.bak
+  if [ -f $data_dir/$mysql_file ]; then
+    sed -i'.bak' -e "s/'gridappsd'@'localhost'/'gridappsd'@'%'/g" $data_dir/$mysql_file
+    # clean up 
+    rm $data_dir/${mysql_file}.bak
+  fi
 fi
 
 if [ ! -f "$data_dir/ieee8500.xml" ]; then
